fix: rename available time directories and files
