+++ __tests__/domain/applications/theoretical_time/theoretical_time_notify_interactor.spec.ts
@@ -1,7 +1,7 @@
 import { UserSetting } from '../../../../src/domain/models/user_setting/user_setting';
 import { UserSettingRepository } from '../../../../src/infrastructure/user_settings/user_setting_repository';
 import { NotifyPresenter } from '../../../../src/app/presenters/notify/notify_presenter';
-import { AvailableTimeNotifyInteractor } from '../../../../src/domain/applications/available_time/available_time_notify_interactor';
+import { AvailableTimeNotifyInteractor } from '../../../../src/domain/applications/theoretical_time/theoretical_time_notify_interactor';
 
 describe('AvailableTimeNotifyInteractor', () => {
   SpreadsheetApp.openById = jest.fn(() => ({
