+++ src/app/controllers/theoretical_time/notify/theoretical_time_notify_controller.ts
@@ -1,4 +1,4 @@
-import { AvailableTimeNotifyUseCaseInterface } from '../../../../../use_case/available_time/notify/available_time_notify_use_case_interface';
+import { AvailableTimeNotifyUseCaseInterface } from '../../../../use_case/theoretical_time/notify/theoretical_time_notify_use_case_interface';
 
 export class AvailableTimeNotifyController {
   constructor(private readonly availableTimeNotifyUseCase: AvailableTimeNotifyUseCaseInterface) {}
