+++ src/google_apps_script/functions/calculate_theoretical_time.ts
@@ -1,8 +1,8 @@
-import { AvailableTimeCalculateInteractor } from '../../domain/applications/available_time/available_time_calculate_interactor';
-import { AvailableTimeRepository } from '../../infrastructure/available_times/available_time_repository';
+import { AvailableTimeCalculateInteractor } from '../../domain/applications/theoretical_time/theoretical_time_calculate_interactor';
+import { AvailableTimeRepository } from '../../infrastructure/theoretical_times/theoretical_time_repository';
 import { IsoWeekRepository } from '../../infrastructure/iso_weeks/iso_week_repository';
 import { UserSettingRepository } from '../../infrastructure/user_settings/user_setting_repository';
-import { AvailableTimeCalculateController } from '../../app/controllers/available_time/calculate/available_time_calculate_controller';
+import { AvailableTimeCalculateController } from '../../app/controllers/theoretical_time/calculate/theoretical_time_calculate_controller';
 
 function calculateAvailableTime(): void {
   const availableTimeRepository = new AvailableTimeRepository();
