+++ src/google_apps_script/functions/notify_theoretical_time.ts
@@ -1,6 +1,6 @@
-import { AvailableTimeNotifyInteractor } from '../../domain/applications/available_time/available_time_notify_interactor';
+import { AvailableTimeNotifyInteractor } from '../../domain/applications/theoretical_time/theoretical_time_notify_interactor';
 import { UserSettingRepository } from '../../infrastructure/user_settings/user_setting_repository';
-import { AvailableTimeNotifyController } from '../../app/controllers/available_time/notify/available_time_notify_controller';
+import { AvailableTimeNotifyController } from '../../app/controllers/theoretical_time/notify/theoretical_time_notify_controller';
 import { NotifyPresenter } from '../../app/presenters/notify/notify_presenter';
 
 function notifyAvailableTime(): void {
