+++ src/use_case/theoretical_time/calculate/theoretical_time_calculate_input_data.ts
@@ -1,4 +1,4 @@
-import { AvailableTimeInputDataBase } from '../available_time_input_data_base';
+import { AvailableTimeInputDataBase } from '../theoretical_time_input_data_base';
 import CalendarEvent = GoogleAppsScript.Calendar.CalendarEvent;
 
 export class AvailableTimeCalculateInputData extends AvailableTimeInputDataBase {
