+++ src/use_case/theoretical_time/notify/theoretical_time_notify_use_case_interface.ts
