+++ src/use_case/theoretical_time/theoretical_time_input_data_base.ts
@@ -1,4 +1,4 @@
-import { Event } from '../../domain/models/available_time/available_time';
+import { Event } from '../../domain/models/theoretical_time/theoretical_time';
 import CalendarEvent = GoogleAppsScript.Calendar.CalendarEvent;
 
 export abstract class AvailableTimeInputDataBase {
